requirements.txt
numpy~=1.19.5
setuptools~=52.0.0
matplotlib~=3.3.2
seaborn~=0.11.1
liac-arff~=2.5.0
sklearn~=0.24.1